// src/interfaces/Paciente.ts
export type Paciente = {
  dh_create_at: Date;
	dh_update_at: Date;
	ds_endereco: string;
	dt_nascimento: Date;
	no_bairro: string;
	no_cidade: string;
	no_email: string;
	no_mae: string;
	no_paciente: string;
	no_pai: string;
	no_responsavel: string;
	no_social: string;
	nu_cep: number;
	nu_cpf: number;
	nu_create: number;
	nu_escolaridade: number;
	nu_estado_civil: number;
	nu_etnia: number;
	nu_paciente: number;
	nu_rg: string;
	nu_situacao_ocupacional: number;
	nu_telefone1: number;
	nu_telefone2: number;
	nu_update: number;
	nu_user: 321;
	sg_sexo: string;
	sg_uf: string;
};